Internal quality cleanup: pub(crate) visibility, find/mod.rs split, TaskInfo.status char
- Narrow 7 hyalo-core pub items to pub(crate) (atomic_write, FilterOp,
  scan_reader_multi, strip helpers, test-only task functions)
- Extract 1,469 lines of find tests to find/tests.rs (mod.rs: 1962 → 491 lines)
- Change TaskInfo/FindTaskInfo status field from String to char
